refactor: Small idiom cleanups (move vec, integer ilog10, newtyped preamble addr)

+++ crates/plotnik-lib/src/bytecode/module/load.rs
@@ -410,7 +410,7 @@ impl Module {
         for i in 0..entrypoints.len() {
             let invalid = || ModuleError::InvalidEntrypoint(i);
             let ep = entrypoints.get(i);
-            let target = ep.target();
+            let target = u16::from(ep.target());
 
             if target >= steps {
                 return Err(invalid());
